feat: refactor and move things to their own files
